🎄Happy New Year🎄 Completed: Ui, ViewModels, Management queue Refactored: TrackRepositoryImpl.kt, fixed parse logic, Navigation

+++ app/src/main/java/com/raaveinm/chirro/data/database/FavTrackInfo.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class FavTrackInfo(
-    @PrimaryKey(autoGenerate = false) val id: Int,
+    @PrimaryKey(autoGenerate = false) val id: Long,
     @ColumnInfo val uri: String,
 )
