+++ app/src/main/java/com/raaveinm/chirro/data/database/TrackDao.kt
@@ -12,5 +12,6 @@ interface TrackDao {
     @Insert (onConflict = OnConflictStrategy.IGNORE) suspend fun insertTrack(track: FavTrackInfo)
     @Delete suspend fun deleteTrack(track: FavTrackInfo)
     @Query("SELECT id FROM FavTrackInfo") fun getFavoriteIds(): Flow<List<Long>>
-    @Query("SELECT EXISTS(SELECT * FROM FavTrackInfo WHERE id = :id)") suspend fun isTrackFavorite(id: Int): Boolean
+    @Query("SELECT EXISTS(SELECT * FROM FavTrackInfo WHERE id = :id)")
+    suspend fun isTrackFavorite(id: Long): Boolean
 }
